requirements_dev.txt
asyncpg==0.23.0
factory-boy==3.2.0
starlette==0.15.0
tortoise-orm==0.17.5
uvicorn==0.14.0
pytest==5.0.1
pytest-asyncio==0.10.0
